+++ packages/rekapager-bundle/config/services.php
@@ -20,14 +20,8 @@ use Rekalogika\Rekapager\Bundle\Twig\RekapagerExtension;
 use Rekalogika\Rekapager\Bundle\Twig\RekapagerRuntime;
 use Rekalogika\Rekapager\Bundle\Twig\TwigPagerRenderer;
 use Rekalogika\Rekapager\Contracts\PageIdentifierEncoderLocatorInterface;
-use Rekalogika\Rekapager\Keyset\PageIdentifierEncoder\SymfonySerializerKeysetPageIdentifierEncoder;
-use Rekalogika\Rekapager\Offset\OffsetPageIdentifierEncoder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Serializer\Encoder\DecoderInterface;
-use Symfony\Component\Serializer\Encoder\EncoderInterface;
-use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 use function Symfony\Component\DependencyInjection\Loader\Configurator\tagged_locator;
@@ -68,24 +62,6 @@ return static function (ContainerConfigurator $containerConfigurator): void {
                 defaultIndexMethod: 'getIdentifierClass'
             )
         ]);
-
-    if (class_exists(SymfonySerializerKeysetPageIdentifierEncoder::class)) {
-        $services
-            ->set(SymfonySerializerKeysetPageIdentifierEncoder::class)
-            ->args([
-                service(NormalizerInterface::class),
-                service(DenormalizerInterface::class),
-                service(EncoderInterface::class),
-                service(DecoderInterface::class),
-            ])
-            ->tag('rekalogika.rekapager.page_identifier_encoder');
-    }
-
-    if (class_exists(OffsetPageIdentifierEncoder::class)) {
-        $services
-            ->set(OffsetPageIdentifierEncoder::class)
-            ->tag('rekalogika.rekapager.page_identifier_encoder');
-    }
 
     $services
         ->set(