+++ packages/rekapager-bundle/src/RekalogikaRekapagerBundle.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace Rekalogika\Rekapager\Bundle;
 
+use Rekalogika\Rekapager\Bundle\DependencyInjection\RemoveMissingDepsPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class RekalogikaRekapagerBundle extends Bundle
@@ -20,5 +22,12 @@ class RekalogikaRekapagerBundle extends Bundle
     public function getPath(): string
     {
         return \dirname(__DIR__);
+    }
+
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new RemoveMissingDepsPass());
     }
 }